+++ src/main/java/org/opensearch/knn/index/memory/NativeMemoryCacheManager.java
@@ -433,11 +433,16 @@ public class NativeMemoryCacheManager implements Closeable {
                 return result;
             }
         } else {
-            // open graphFile before load
             try (nativeMemoryEntryContext) {
                 String key = nativeMemoryEntryContext.getKey();
-                open(key, nativeMemoryEntryContext);
-                return cache.get(key, nativeMemoryEntryContext::load);
+                // if we already have the allocation we should not open file again as this will cause slowdown in
+                // heavy throughput cases, since open() function do locking while opening and mapping the graph file to
+                // memory.
+                return cache.get(key, () -> {
+                    // open graphFile before load
+                    open(key, nativeMemoryEntryContext);
+                    return nativeMemoryEntryContext.load();
+                });
             }
         }
     }